chore: split up code
